+++ Sources/View/Channel/MessageCell/MessageCellParams/SBUUserMessageCellParams.swift
@@ -15,10 +15,40 @@ public class SBUUserMessageCellParams: SBUBaseMessageCellParams {
     public let useReaction: Bool
     public let withTextView: Bool
     
-    public init(message: UserMessage, hideDateView: Bool, useMessagePosition: Bool, groupPosition: MessageGroupPosition = .none, receiptState: SBUMessageReceiptState = .none, useReaction: Bool = false, withTextView: Bool, isThreadMessage: Bool = false, joinedAt: Int64 = 0, messageOffsetTimestamp: Int64 = 0) {
-        
+    /// The boolead value to indicates that the message cell should hide suggested replies.
+    /// If it's `true`, never show the suggested replies view even the `BaseMessage/ExtendedMessagePayload` has the reply `option` values.
+    /// - Since: 3.11.0
+    public let shouldHideSuggestedReplies: Bool
+    
+    /// The boolead value to indicates that the message cell should hide form type message.
+    /// If it's `true`, never show the form type message view even the `BaseMessage/ExtendedMessagePayload` has the reply `forms` values.
+    /// - Since: 3.11.0
+    public let shouldHideFormTypeMessage: Bool
+
+    /// The form answers are the data for the message cell to redraw each user's form field input.
+    /// - Since: 3.11.0
+    public let formAnswers: [SBUForm.Answer]
+
+    public init(
+        message: UserMessage,
+        hideDateView: Bool,
+        useMessagePosition: Bool,
+        groupPosition: MessageGroupPosition = .none,
+        receiptState: SBUMessageReceiptState = .none,
+        useReaction: Bool = false,
+        withTextView: Bool,
+        isThreadMessage: Bool = false,
+        joinedAt: Int64 = 0,
+        messageOffsetTimestamp: Int64 = 0,
+        shouldHideSuggestedReplies: Bool = true,
+        shouldHideFormTypeMessage: Bool = true,
+        formAnswers: [SBUForm.Answer]? = nil
+    ) {
         self.useReaction = useReaction
         self.withTextView = withTextView
+        self.shouldHideSuggestedReplies = shouldHideSuggestedReplies
+        self.shouldHideFormTypeMessage = shouldHideFormTypeMessage
+        self.formAnswers = formAnswers ?? []
         
         var messagePosition: MessagePosition = .left
         if useMessagePosition {